# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tawjeeh adaptive learning API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /learning/exam-mode:
    get:
      operationId: getExamMode
      tags: [learning]
      summary: Get the active baccalaureate preparation mode
      parameters:
        - name: exam_date
          in: query
          required: false
          schema:
            type: string
            description: Target baccalaureate date in YYYY-MM-DD format
      responses:
        "200":
          description: Active exam preparation mode and high-error concepts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamMode"
...
  /learning/attempts:
    post:
      operationId: recordLearningAttempt
      tags: [learning]
      summary: Record a practical attempt and link errors to a summary concept
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LearningAttemptInput"
      responses:
        "201":
          description: Attempt metrics and any generated remediation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningAttemptResponse"
...
  /quizzes:
    get:
      operationId: listQuizzes
      tags: [quizzes]
      summary: List available quizzes
      parameters:
        - name: exam_date
          in: query
          required: false
          schema:
            type: string
            description: Target baccalaureate date in YYYY-MM-DD format
      responses:
        "200":
          description: Available quizzes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quiz"
...
  /quizzes/{quizId}/attempt:
    post:
      operationId: submitQuizAttempt
      tags: [quizzes]
      summary: Submit a quiz answer set
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizAttemptInput"
      responses:
        "200":
          description: Quiz result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"
        "404":
          description: Quiz not found
components:
  schemas:
...
    ConceptMetric:
      type: object
      required: [lesson_id, lesson_title, concept_id, concept_title, attempts, errors_count, error_rate, last_error_tag]
      properties:
        lesson_id:
          type: string
        lesson_title:
          type: string
        concept_id:
          type: string
        concept_title:
          type: string
        attempts:
          type: integer
        errors_count:
          type: integer
        error_rate:
          type: number
        last_error_tag:
          type: string
...
    LearningAttemptInput:
      type: object
      required: [lesson_id, lesson_title, concept_id, concept_title, error_tag, is_correct]
      properties:
        lesson_id:
          type: string
        lesson_title:
          type: string
        concept_id:
          type: string
        concept_title:
          type: string
        error_tag:
          type: string
        is_correct:
          type: boolean
    LearningAttemptResponse:
      type: object
      required: [attempt_id, metric, remediation]
      properties:
        attempt_id:
          type: integer
        metric:
          $ref: "#/components/schemas/ConceptMetric"
        remediation:
          oneOf:
            - $ref: "#/components/schemas/ScheduleEntry"
            - type: "null"
    ScheduleEntry:
      type: object
      required: [id, scheduled_date, time, duration, title, subject, kind, remediation_label, lesson_id, concept_id, completed, missed, penalty_type, volume_multiplier]
      properties:
        id:
          type: integer
        scheduled_date:
          type: string
        time:
          type: string
        duration:
          type: string
        title:
          type: string
        subject:
          type: string
        kind:
          type: string
        remediation_label:
          type: ["string", "null"]
        lesson_id:
          type: ["string", "null"]
        concept_id:
          type: ["string", "null"]
        completed:
          type: boolean
        missed:
          type: boolean
        penalty_type:
          type: ["string", "null"]
        volume_multiplier:
          type: integer
...
    ExamMode:
      type: object
      required: [mode, label, description, exam_date, days_until, exercise_density, reduce_passive_explanation, error_concepts]
      properties:
        mode:
          type: string
          enum: [standard, pre_exam, error_stack]
        label:
          type: string
        description:
          type: string
        exam_date:
          type: string
        days_until:
          type: integer
        exercise_density:
          type: integer
          minimum: 1
        reduce_passive_explanation:
          type: boolean
        error_concepts:
          type: array
          items:
            $ref: "#/components/schemas/ConceptMetric"
...
    QuizQuestion:
      type: object
      required: [id, prompt, options, concept_id, concept_title, source_node_ids]
      properties:
        id:
          type: string
        prompt:
          type: string
        options:
          type: array
          items:
            type: string
        concept_id:
          type: string
        concept_title:
          type: string
        source_node_ids:
          type: array
          minItems: 1
          items:
            type: string
    Quiz:
      type: object
      required: [id, title, subject, description, duration, questions, status, points, is_high_difficulty, unit_id, score_threshold, mode, exercise_density, reduce_passive_explanation, linked_concepts]
      properties:
        id:
          type: string
        title:
          type: string
        subject:
          type: string
        description:
          type: string
        duration:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
        status:
          type: string
        points:
          type: number
        is_high_difficulty:
          type: boolean
        unit_id:
          type: string
        score_threshold:
          type: number
        mode:
          type: string
          enum: [standard, pre_exam, error_stack]
        exercise_density:
          type: integer
          minimum: 1
        reduce_passive_explanation:
          type: boolean
        linked_concepts:
          type: array
          items:
            type: string
    QuizAttemptInput:
      type: object
      required: [answers]
      properties:
        answers:
          type: object
          additionalProperties:
            type: string
        exam_date:
          type: string
          description: Target baccalaureate date used to resolve adaptive exam content
    QuizResult:
      type: object
      required: [quiz_id, score, correct, total, message, points_earned, attempt_id, passed, is_high_difficulty, mode, linked_concepts]
      properties:
        quiz_id:
          type: string
        score:
          type: number
        correct:
          type: number
        total:
          type: number
        message:
          type: string
        points_earned:
          type: number
        attempt_id:
          type: integer
        passed:
          type: boolean
        is_high_difficulty:
          type: boolean
        mode:
          type: string
          enum: [standard, pre_exam, error_stack]
        linked_concepts:
          type: array
          items:
            type: string
...